fix the price field in product model

+++ product/models.py
@@ -1,3 +1,20 @@
+# from django.db import models
+#
+# from product.querysets import ProductQuerySet
+#
+#
+# class Product(models.Model):
+#     name = models.CharField(max_length=100)
+#     price = models.FloatField()
+#     description = models.TextField()
+#     stock = models.IntegerField(default=0)
+#
+#     objects = ProductQuerySet.as_manager()
+#
+#     def __str__(self):
+#         return self.name
+
+
 from django.db import models
 
 from product.querysets import ProductQuerySet
@@ -5,7 +22,7 @@ from product.querysets import ProductQuerySet
 
 class Product(models.Model):
     name = models.CharField(max_length=100)
-    price = models.IntegerField()
+    price = models.FloatField()
     description = models.TextField()
     stock = models.IntegerField(default=0)
